Fix 3 hidden memory spikes: wavelet chunking, ranked_signals f32, size_neutralize f32
Wavelet FFT (features.py):
- Was allocating 30 GB complex128 array from full-matrix rfft on float64
- Now processes in 500-ticker chunks with float32 input → complex64 output
- Peak per chunk: 1.4 GB (was 30 GB). Intermediates del'd after each chunk

ranked_signals float32 (features.py):
- cross_sectional_rank().rank(pct=True) returns float64 by default
- Added .astype(np.float32) to all 24 ranked signal DataFrames
- Saves 1.6 GB (3.2 GB → 1.6 GB)

size_neutralize float32 (model.py):
- OLS computation used dtype=float (=float64) for numpy arrays
- Changed to dtype=np.float32. Saves ~0.5 GB transient per _sn variant

Combined savings: ~32 GB peak reduction (wavelet) + 2.1 GB steady-state.
Expected peak with all features: ~22-24 GB (fits in 32 GB without swap).

+++ src/model.py
@@ -111,8 +111,8 @@ def size_neutralize(
     lm = log_mcap_df.reindex(index=feature_df.index, columns=feature_df.columns)
     out = feature_df.copy()
 
-    x_arr = lm.to_numpy(dtype=float)
-    y_arr = feature_df.to_numpy(dtype=float)
+    x_arr = lm.to_numpy(dtype=np.float32)
+    y_arr = feature_df.to_numpy(dtype=np.float32)
     res = y_arr.copy()
 
     for i in range(y_arr.shape[0]):
